feat(route): recover kaopu.news route
