+++ lib/routes/kaopunews/namespace.ts
@@ -1,0 +1,6 @@
+import type { Namespace } from '@/types';
+
+export const namespace: Namespace = {
+    name: '靠谱新闻',
+    url: 'https://kaopu.news',
+};
